Pull out duplicate readiness polling code
